chore: add assigning folderId to new group
